requirements.txt
asgiref==3.11.1
Django==5.2.12
gunicorn==25.3.0
packaging==26.0
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.12.0
markdown==3.7
requests==2.31.0
feedparser==6.0.11
beautifulsoup4==4.12.3
google-genai==1.72.0